Pushing in the front-end HTML file for test.sol
Quite a lot left to learn. I wonder what I am doing wrong or
is it just the cdn and jquery links that are malfunctioning. Maybe I should try and build this on the newer versions of perhaps on truffle

`jQuery.Deferred exception: web3 is not defined ReferenceError: web3 is not defined
    at HTMLDocument.<anonymous> (file:///C:/CS_07/Software_Projects/BlockChain_Projects/Dapp-Dev/frontEnd_test/index.html:21:33)
    at mightThrow (https://code.jquery.com/jquery-3.6.0.js:3766:29)
    at process (https://code.jquery.com/jquery-3.6.0.js:3834:12) undefined`